Enable CodeGenHashStability tests on Linux/Unix platforms
On my local machine the CodeGenHashStability test takes 14 seconds to
run. While that is longer than the entire rest of the test suite, it
makes the full test suite run only take 16 seconds, which is still
reasonable.

We should enable this test case by default and evaluate ways to better
parallelize this test so that so that we can run it more efficiently.