Add tests for multi-dim arrays and unbounded arrays (#5532)
I added at test for unbounded arrays. This case works, so no code
changes were required.

I also added a test for multi-dimensional arrays. This case is not
handled allowed in Vulkan, so I chose to issue an error. If we need to
support it, we could try improve it later by transforming the array into
a single dimensional array.

See https://gitlab.khronos.org/spirv/SPIR-V/-/issues/743 and
https://registry.khronos.org/vulkan/specs/1.3-extensions/html/vkspec.html#interfaces-resources-setandbinding.