update dndxc to support latest shader models (#5565)
dndxc has a dropdown to indicate targets. It has to be updated to latest
shading models to be able to use them. This adds 6.7 and 6.8 variants
for all stages as well as adding the mesh and amplification shader stages 
and removing the invalid lib_6_1 and lib_6_2 stages

Included 6.8 even though it's not in main yet. Best to be prepared