Indicate extra metadata with unknown nodio tag
When unrecognized metadata is encountered, it needs to be recorded so
that validation will fail. Additionally, this removes the assert that
was there so that the code can be tested. Asserts for code that can be
encountered through potential input isn't ideal anyway.

Adds testing for the same which involves introducing dxv testing into
the llvm test directory requiring enabling .ll tests for that
subdirectory and enabling lit to understand %dxv.

Incidentally fixes some comments in unrelated workgraph tests that had
outdated values for some of the metadata that might lead to confusion

Fixes #5609