Remove unneeded Sema equal param special case
This was added to address a special case where OutputComplete was
expecting a templated type, but got the wrong one and gave an unhelpful
"no matching function" error instead of the error more specific to the
requirements of OutputComplete.

Since OutputComplete no longer takes this as a parameter and is instead
a method of the object, the hack is not needed.

Note that this is tested by case e in
case127_outputcomplete_errors.hlsl:
[Shader("node")]
[NodeLaunch("Broadcasting")]
void node127_e([MaxRecords(5)] EmptyNodeInput emptyNodeInput)
{
  emptyNodeInput.OutputComplete();
}

part of #5364