Correctly use halfword-aligned byte address
All loads and stores for ByteAddressBuffer were being aligned to the
nearest word. This is incorrect behavior for 16-bit types. In order to
implement this, I had to move some of the logic for calculating offsets
from compile time to runtime. For loads and stores of 32 or 64-bit types
with a literal byte address, the optimizer will be able to reduce this
down to nearly the same code as was being output before.

For 16-bit types, stores are slightly more convoluted (we need to load
the old value, shift and mask it, and then combine it with the new
value, in order to avoid clobbering any existing value). The optimizer
isn't currently able to optimize away unnecessary shifts and masks, but
my intuition is that it should be possible, in case we want to optimize
this more.

Fixes #5475