MergeGepUse for Lowered UDT to avoid invalid dxil on -Od
Without optimizations, GEP chains may be used to access multi-dim arrays within structures.
These arrays are legal within the DXIL structure, but not allowed as an intermediate pointer returned from a GEP.

This change runs MergeGepUse after lowering UDT types to merge and eliminate GEPs that return multi-dim array pointers.

A side-effect of running MergeGepUse is that GEPs are marked as inbounds.