Require validator ver 1.7 for link without version (#5629)
The container builder determines whether to include the version in the
compiled shader library depending on the validator version. If the
version is below 1.8, it won't include it and any linking is allowed. If
it is 1.8 or above, the version will prevent linking with a
non-versioned compile.

RunLinkWithTempReg assembled one library shader and compiled another. As
such, one had a version and the other didn't. By forcing the compile to
use an earlier validator version, the link is again allowed

Followup to #5378