Avoid extra dependencies in the DxilRuntimeReflection header. NFC
including `WinIncludes.h` here pulls in a bunch of dependencies like
ATL even though we really don't need those in this header. Given that
the only things we're using that include for is to use LPCWSTR and
LPCVOID spellings in a way that works on both windows and unix, just
expand those macros to their definitions instead.