Consolidate EmitInstrNote into common function (#5615)
EmitInstrNote reused most of the code from EmitInstrErrorMsg. This
consolidates both functions into one that shares most of the code. In
addition, it no longer makes the note look like an additional error in
the output and adds testing for that note. Relevant tests are modified
to include the notes

Fixes #5613