Handle log10 with float16_t parameter (#5659)
When the spirv backend sees the log10 intrinsic, it will generate a
multiple with a constant. The type of that constant is hard coded to be
a float. However, if the type of the parameter is not a float, there is
a type mismatch on the multiple instruction.

This commit changes that so the type of the contanst matches the scalar
type of the parameter. That is, it will be a type T if the parameter is
type T, a vector of type T, or a matrix of type T.

Fixes #5608