[NFC] Fix ubsan failure
This variable was never initialized so it was always whatever garbage
bits were left there. I've defaulted it to `true` because a bunch of
tests fail if it is set to `false`, but all tests pass if it is true.

It would be great if someone with more SPIR-V experience could take a
look at this.