Disallow multiple shader attributes from existing on a single declaration (#5679)
There was a recent spec change on work-graphs that disallows the
existence of node + compute attributes on shader declarations. Because
of this old behavior, a lot of code worked around the possibility that 2
shader attributes could exist for one declaration. With this old
behavior disallowed, we no longer need to account for this possibility.
The PR now adds a diagnostic preventing multiple attributes from
existing. It will also change all the tests that worked with shaders
with multiple shader attributes, and either remove them, or update them
so that only one attribute is used.
Fixes #5601