Improve diagnostics for objects in node records
The diagnostics produced for the use of objects within node
records were rather generic and obscure.
Improved the diagnosis of such issues by adding more specific
diagnostics, and moved exisiting verifier tests from a catch-all
test case to a specific test for invalid node record types and
added more test cases.

It transpired that templated types within a node record are
problematic - they are not currently diagnosed correctly and
produce a zero size record diagnostic, and if this is bypassed
they hit asserts during CodeGen. So as a lesser evil, an error
diagnostic has been added to disallow templated types for now.