[NFC] clang-format DXC sources
This change applies clang-format to all of the sources added in DXC that
were not part of the original LLVM/Clang 3.7 release when DXC forked.
This provides consistent code formatting across the codebase. All future
changes are expected to conform to clang-format which will be enforced
by a GitHub action added in PR #5617.

Fixes #5591.