Validate patchconst functions for GS/update test
When compiled as libraries, patch constant functions didn't get
validated for use of groupshared like their entry point counterparts
did. This adds that validation step and produced an equivalent error.

In addition, the test that should have caught this was missing checks
for that, but also had 1/4 of the checks that it needed because the
values were broken out into scalars for the purposes of validation so
there is one error for every scalar value.

Updates that same test with a node entry point to verify that
groupshared memory can be used there. This also requires updating the
code to use payload access qualifiers for the raytracing elements.

Fixes #5748