[NFC] Fix clang-formatting generated sources
Clang-format determines if a header is the "main" header by looking at
the header name without the last extension and matching it against the
source file (with some extra handling for cuda files).

The automation generating temporary files with a `.tmp` suffix tripped
this up so it wasn't matching the "main" header when sorting includes.

This corrects the issue by writing temporaries to a tmp directory
instead of munging the suffix.