Add a clang-format error to force workflow failure (#5757)
This change _should_ fix the permissions issues and allow the PR format
checker to post comments into PRs on failure. The comments should
include the first 50 lines of the diff that clang-format produces and an
error with a link to the full log.

Unfortunately this can't really be tested until it is merged. The
`pull_request_target` event only occurs once the action is in the target
branch.

Parts of this change have been tested separately, but they haven't all
been tested together.