Fix the clang-format workflow (#5761)
There were a few bugs that were difficult to reproduce. The pull request
branch ref needs to be fetched and checked out manually in order to get
the PR diff to construct. This change also fixes formatting to get a
nice diff printed of the PR output.

See an example of the error output this produces here:

https://github.com/llvm-beanz/DirectXShaderCompiler/pull/2