return E_NOINTERFACE for invalid iid on create; support IUnknown
Added support to IUnknown on create because it could be useful to have common code path for either library or shader reflection.
Without this, this change could have some regression risk, since this path would have worked before.
Since the distinction is determined by what's in the reflection blob, the specific interface isn't required here.
For shader reflection, this will set the interface version to the latest (D3D12).