WinPIX: Find correct filename for subroutines in included headers within namespaces (#5776)
From a customer bug report, the previous code would skip over a
DISubProgram scope if that scope was contained in a DINamespace (as in
the first test case herein, wherein an include file is enclosed in a
namespace). This would result in DXIL being attributed to the includer
file instead of the included file.

(cherry picked from commit 5adc7bcf678a39d82be532cb1c3615195e3f420d)