Add test for FieldAnnotationVectorSize and similar metadata tags (#5786)
This test checks that a vector size annotation has been added for vector
field. The annotation is embedded in metadata used for reflection. The
test is using the `-Qkeep_reflect_in_dxil` flag to keep the metadata in
the module to allow metadata check in the test, otherwise it gets
stripped from the main module.

The metadata is embedded only for DXIL version 1.8 and higher.

Similar metadata checks added for matrix and scalar annotations for
completeness.

Fixes #5697