Make member expressions of bitfields emit as an RValue instead of LValue (#5740)
Previously, when emitting a vector splat type as a cast LValue, the only
valid types that could be casted were simple and Extended vector element
types. If, (in this case), a bitfield type was detected, an assert would
be triggered causing an internal compiler error. This change adds cases
to handle all but one of the possible expected types to cast from. A
test was added specifically for the bitfield conversion process. The
regression test uses the shadaer supplied in the original issue, and
this PR now successfully compiles that shader.
Fixes #4758