Improve library profile diagnostics (#5789)
There are several diagnostics that are missing when a function
declaration that targets a specific shader stage is missing a necessary
attribute. For example, if a hull shader function declaration does not
have the patchconstantfunc attribute, there is no diagnosis of the
missing attribute in sema. This PR addresses this diagnostic hole by
filling in some functions that check that all required attributes exist
for a given shader stage.
This PR partially addresses #5747