Move NodeMaxDispatchGrid diagnostic into Sema (#5806)
The diagnostic that is emitted when `NodeMaxDispatchGrid` attribute is
used on a structure that has no members with the `SV_DispatchGrid`
semantic is implemented in `CGHLSLMS.cpp`. This location isn't ideal.
This PR is moving that diagnostic to an earlier compile stage into
`SemaHLSL.cpp`. That way the check can happen earlier and will be
emitted with `dxc -verify`.

New test cases for the `SV_DispatchGrid` diagnostic were added to
`VerifierTests` and corresponding tests in codegen removed.

Since the diagnostics was implemented in codegen, many other verifier
tests did not hit it and were using `NodeMaxDispatchGrid` without a
valid node input record. These tests have been adjusted on a
case-by-case basis either by adding a field with `SV_DispatchGrid`,
changing the attribute to `NodeDispatchGrid` or adding the error to
expected output.

Fixes #5775