Add nullptr checks to handle undefined arrays
When a variable declaration is not found, an error is emitted, and a null value returned. Not all function know what to do with that null value. I fix this by adding a few extra nullptr checks so that the compiler can unwind the stack and return gracefully.

Fixes [SPIR-V] `-Wundefined-internal` SIGSEGVs the compiler when `-spirv` output #5821