ATOProgOffset test fix
The test was broken by commit 1dac9d8f518b429c3b9f95aad9fcb58d1e0d4591 that was eliminating clang compiler warnings in HLSL Execution Tests, specifically casting signed shift to unsigned. In this particular case the signed shift was intentional. This PR is changing the unsigned shift to signed and disabling warning for this operation for clang.