Make hctbuild cmake gen consistent with asserts
A slightly complex series of events causes invocations of hctbuild after
the first to enable asserts in release builds. This sets the build type
even when it isn't needed to avoid it.

Because the build type was not being set when generating for visual
studio in hctbuild, when generating the cmake files using CMakeLists,
it would produce an alarming message saying it was
defaulting to debug, but, in fact, since it didn't care for the purposes
of sln file generation, that wasn't actually a problem, just confusing.
It did set a variable to enable asserts, but didn't do this for release
variant builds

The real problem occurs if hctbuild is invoked again, at that point,
instead of defaulting to Debug, it leaves the build type empty because of
another variable set in the cache.

Additionally, since the first invocation enabled asserts, that also
comes from the cache and so code that strips NDEBUG from all build
variants is invoked because the build type is, in fact, not debug, it's
an empty string.