Make DxilSignatureElement::m_SemanticName canonical for system values (#5889)
When DxilModule metadata is serialized, semantic strings for system
values are canonicalized. If these strings are used anywhere, this can
lead to a difference in behavior between an original DxilModule that has
not undergone serialization and one that has been deserialized.

This change canonicalizes the string as soon as a DxilSignatureElement
is created, thereby removing the possibility of accidentally relying on
the non-canonical string for a system value.