[SPIR-V] Correctly handle subscript overload in access chain (#5896)
In order to handle types such as `RWStructuredBuffer`, which has the
method `operator[]()`, `operator[]` is treated as a special case by our
access chain creation code - it's treated similarly to
`ArraySubscriptExpr`. This has the unfortunate side effect that
user-defined `operator[]` overloads are not called in all cases.

To fix this, this PR adds a check for whether a type is user-defined
before handling `operator[]` in `collectArrayStructIndices`.

Part of the fix for #5638.