Do not merge geps when an addrspacecast changes the underlying type.
We do not handle an addrspacecast that changes the shape of the
underlying type when merging geps (e.g. array to non-array). Instead of
trying to handle some cases we just avoid trying to merge geps in all
cases where the underlying type changes shape. This causes the
multi-dim-one-dim pass to crash because it expects the geps to be
merged. I do not think we can hit this case from hlsl, but it may be
possible.