[SPIR-V] Remove invalid CHECK-COUNTs
CHECK-COUNT-# was not yet implemented in the version of FileCheck used
in DXC, so these lines were not checked at all. Switching them to CHECKs
caused the tests to fail, so they're now replaced with a combination of
a CHECK and CHECK-NOT that pass.

This was caught by the downstream test runner that uses a newer version
of FileCheck which does check CHECK-COUNTs.