Add missing '-' for pass in ExtractIRForPassTest.py
The default generated RUN line generated by ExtractIRForPassTest.py adds the
desired pass to the options, but forgot to add the '-', so it would fail
with difficult to diagnose error in dxopt during argument parsing.

This adds the missing '-' so the RUN line doesn't always have to be fixed in
this way everytime the script is used.