Make DxilRemoveDeadBlocks init DxilModule, fix test (#5919)
DxilRemoveDeadBlocks assumed it could run with an uninitialized
DxilModule, because the test case had been captured from live state
without the needed DXIL metadata. DxilRemoveDeadBlocks relies on
DxilValueCache, which relies on code in DxilSimplify, which assumes we
have a DxilModule if there are dxil ops. Without an initialized
DxilModule with the correct low-precision mode, cached ops and types may
be wrong.

This change:
- modifies DxilRemoveDeadBlocks to remove bSkipInit
- updates the test to contain the metadata needed to initialize
DxilModule