[CI] add workflow to apply clang-format diff
When hit clang-format error in a pull request, edit the comment with diff.
Add @apply-format at the beginning of the comment and Update comment.
New added "Apply code format" workflow will be triggered.
It will checkout the code from compare branch of the pullrequest, apply the diff,
then commit and push it to the compare branch to fix the format issue.