Simplify CXTranslationUnit Creation
The original code called a function whose expected behavior was to create a TU via an out param.
The preprocessor changes for DXC disabled that code path causing an eroneous clang tidy warning.
Since that function didn't do anything it is equally correct to just not call it. This change does that.