[SPIR-V] Fix crash with GetAttributeAtVertex
The code assumed all the types we handled had an AST type.
But when the struct is created in the SPIR-V backend, there are
no AST type left to represent it.
In such case, the lowering step should handle the interpolation
attributes, so I think we shouldn't have to do anything.

Signed-off-by: Nathan Gauër <brioche@google.com>