[Diag] Warn instead of error on invalid numthreads
If an invalid numthreads attribute isn't used, erroring out on it is a
breaking change that might be annoying to deal with. Instead we make
invalid numthreads a warning and drop the attribute - we can rely on
the existing diagnostics for missing numthreads to error out in cases
that aren't recoverable.

Reviewers: bob80905, damyanp, llvm-beanz

Reviewed By: llvm-beanz, bob80905, damyanp

Pull Request: https://github.com/microsoft/DirectXShaderCompiler/pull/6023