[SPIR-V] Fix interlock output parameter validation
Interlock* functions takes output parameters. Because HLSL had no
real concept of references, this is allowed:

```
  int value;
  takeAnOutput(value);
```

This weirdness makes using `isLValue()` to validate output parameters
a bit tricky. If a function returns a local array, and we take the
subscript of this r-value, the r-value is now marked as l-value, but
this shouldn't be allowed.

This forces us to walk the expression, and add heuristics to determine
if the output parameter is valid or not. Not pretty, but I think this is
required until HLSL gets proper references.

Signed-off-by: Nathan Gauër <brioche@google.com>