[Diag] Reword `err_hlsl_missing_attr`
Instead of "must have the X attribute" we say "must have a valid X
attribute". This should make things clearer in cases where there was
something wrong with the attribute and we've dropped it, like we do if
`numthreads` is specified incorrectly.

Reviewers: bob80905, llvm-beanz, pow2clk

Reviewed By: bob80905, pow2clk, llvm-beanz

Pull Request: https://github.com/microsoft/DirectXShaderCompiler/pull/6024