Only diagnose exported library function declarations (#6009)
Previously, diagnosing the translation unit would loop through every
function declaration found in a library shader. This behavior is
undesired, as there are some functions that are not intended to be
exported, and thus shouldn't emit diagnostics (maybe the library
compilation target is for a shader model that wouldn't mesh well with
some other function declarations in the translation unit). This PR
adjusts the diagnostic emission code so that only exported functions are
diagnosed.
Fixes https://github.com/microsoft/DirectXShaderCompiler/issues/5857