[SPIRV] allow location index to be > than 64 (#6015)
There was an arbitrary limit to set the maximum number of locations to
64.
This commit allows the set to be resized, to allow arbitrarily large
location indices.

Curiously, one test, vk.location.large.hlsl started to fail. This test
assumed 123456 was an invalid location. From the SPIR-V spec, I don't
think this is true, as it only days Location should be an arbitraty
number.
In addition, it seems like this test was triggering an out-of-bound
access, which is now solved. Only drawback: we now end-up with a 123456
sized bitset.
I believe this is OK, as WHY would someone only require this index.

Fixes #3735

---------

Signed-off-by: Nathan Gauër <brioche@google.com>