[SPIR-V] Enable vk::ext_extension and vk::ext_capability for variables (#6052)
`ext_extension` and `ext_capability` can only be added to functions at
the moment. This PR enables adding them to variable declarations, so
that using the variable results in the extensions and capabilities being
added to the module. This is useful for the builtin variable syntax
proposed in https://github.com/microsoft/hlsl-specs/pull/129.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>