Fix segfault when compiling hlsl intrinsics with variadic argu… (#6012)
For hlsl intrinsics with variadic argument lists (in my case: printf),
the number of param modifiers and function arguments did not match,
resulting in segfaults when creating a FunctionProtoType. The presumed
oversight: The paramsMods was made twice as large as intended, by first
resizing to the expected number of arguments, and then pushing back the
same number of arguments again. A fix that seems to resolve this issue
comes with this PR. Please have a look.

Edit: The suggested modifications do not feel as if they are the proper
way to handle modParams initialization. They only defuse the issue I had
with printf. If you think this issue needs a better solution and/or a
proper bug report before suggesting a PR, please let me know - and
please apologize in case I opened up a PR too early.