Fixed an OOB that potentially causes a crash (#6079)
While emitting diagnostic notes about conversions, the code for checking
`OutConversions` was not included in the loop. This caused an OOB when
`I >= NumConversions`. Normally, this does not affect anything (other
than not emitting the diagnostics for `OutConversions`), since most of
the time `OutConversions[I].isBad()` happens to return false. However,
in some cases when `OutConversions[I].isBad()` is true, calling
`DiagnoseBadConversion()` on the invalid entry will crash.