[SPIRV] Emit OpLine for internally generated SPIR-V wrapper function (#6074)
This PR improves the experience for profilers and debuggers that consume
SPIR-V produced by dxc. When debugging information is output,
additionally note the beginning of the internally generated wrapper
function. This will assign all parameter setup to the entry point
declaration, providing context to the user of a shader profiler or
debugger.