Refactor create stage vars (#6059)
The `createStageVars` function is very large, and difficult to fix. In
support of two bug fixes, I want to refactor the function so try to make
the code cleaner.

This change will try to keep the exact same functionality. I believe
their could be further simplification after this if needed.